Check and fill size_cell from ODIM
git-svn-id: https://svn.smr.arpa.emr.it/radar/trunk@1608 5e904517-b230-484e-af0d-ad602924471e

+++ include/cum_bac.h
@@ -67,7 +67,7 @@ class Volume
 {
 public:
     time_t acq_date;
-    float size_cell;
+    double size_cell;
     bool declutter_rsp; // ?
 
     Volume()
